+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 
 ## [Unreleased]
 ### Fixed
+- [`prefer-default-export`] handles flow `export type` ([#484])
 - [`prefer-default-export`] handles re-exported default exports ([#609])
 - Fix crash when using `newline-after-import` with decorators ([#592])
 - Properly report `newline-after-import` when next line is a decorator